Compilable/spetializable without errors
-- refactored - filesystem added
-- before testing basic annealing population

+++ probability.h
@@ -7,6 +7,6 @@ class rnd{
 public:
 	static unsigned int lrand(){return 0;}
 public:
-	static unsigned int Init(){srand(time(NULL)); return RAND_MAX;}
+	static unsigned int Init(){srand((unsigned)time(NULL)); return RAND_MAX;}
 	static unsigned int U(unsigned int a, unsigned int b){return 0;}
 };
